chore: refactor cm language files
