+++ frontend/src/core/codemirror/language/embedded/latex.ts
